refactor(dashboard): modularize UI javascript code
- Create a file for every individual piece of js functionality.
- This allows for reuse in other UI files.

[Delivers #166315493]
