+++ UI/authentication.html
@@ -90,6 +90,6 @@
         </div>
 
         <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-        <script src="./js/authentication.js"></script>        
+        <script type="module" src="./js/authentication.js"></script>        
     </body>
 </html>
