+++ UI/dashboard.html
@@ -50,7 +50,10 @@
                                     <ul id="sub_menu">
                                         <div class="submenu_username">
                                             <a href="./dashboard.html">
-                                                <div><span class="capitalize firstname"></span> <span class="capitalize lastname"></span></div>
+                                                <div>
+                                                    <span id="sub_menu_firstname" class="capitalize firstname"></span>
+                                                    <span id="sub_menu_lastname" class="capitalize lastname"></span>
+                                                </div>
                                             </a>
                                         </div>
                                         <hr class="submenu_hr">
@@ -74,7 +77,11 @@
                 
                 <section id="dashboard">
                     <div class="container">
-                        <h2 class="page_heading"><span class='capitalize firstname'></span> <span class='capitalize lastname'></span> <i class="fas fa-check-circle"></i></h2>
+                        <h2 class="page_heading">
+                            <span id="dash_firstname" class='capitalize firstname'></span>
+                            <span id="dash_lastname" class='capitalize lastname'></span>
+                            <i class="fas fa-check-circle"></i>
+                        </h2>
 
                         <hr class="section_divider">
 
@@ -86,38 +93,31 @@
                         
                         <div class="user_information">
                             <div class="user_card card">
-                                <div
-                                    id="dp"
-                                    class="user_photo"
-                                    style="background-image: url('https://s3.eu-west-2.amazonaws.com/quick-credit/profile_photos/1')"
-                                >
-                                <svg
-                                    id="profile_pix"
-                                    aria-hidden="true"
-                                    focusable="false"
-                                    data-prefix="fas"
-                                    data-icon="camera"
-                                    class="svg-inline--fa fa-camera fa-w-16"
-                                    role="img"
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    viewBox="0 0 512 512">
-                                    <path
-                                        d="M512 144v288c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V144c0-26.5 21.5-48 48-48h88l12.3-32.9c7-18.7 24.9-31.1 44.9-31.1h125.5c20 0 37.9 12.4 44.9 31.1L376 96h88c26.5 0 48 21.5 48 48zM376 288c0-66.2-53.8-120-120-120s-120 53.8-120 120 53.8 120 120 120 120-53.8 120-120zm-32 0c0 48.5-39.5 88-88 88s-88-39.5-88-88 39.5-88 88-88 88 39.5 88 88z">
-                                    </path>
-                                </svg>
-
-                                <!-- <img
-                                    title="Click to change"
-                                    id="profile_pix"
-                                    src="https://s3.eu-west-2.amazonaws.com/quick-credit/profile_photos/1"
-                                    alt="Profile picture"
-                                > -->
+                                <div id="dp" class="user_photo">
+                                    <svg
+                                        id="profile_pix"
+                                        aria-hidden="true"
+                                        focusable="false"
+                                        data-prefix="fas"
+                                        data-icon="camera"
+                                        class="svg-inline--fa fa-camera fa-w-16"
+                                        role="img"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                        viewBox="0 0 512 512">
+                                        <path
+                                            d="M512 144v288c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V144c0-26.5 21.5-48 48-48h88l12.3-32.9c7-18.7 24.9-31.1 44.9-31.1h125.5c20 0 37.9 12.4 44.9 31.1L376 96h88c26.5 0 48 21.5 48 48zM376 288c0-66.2-53.8-120-120-120s-120 53.8-120 120 53.8 120 120 120 120-53.8 120-120zm-32 0c0 48.5-39.5 88-88 88s-88-39.5-88-88 39.5-88 88-88 88 39.5 88 88z">
+                                        </path>
+                                    </svg>
                                     <input type="file" id="img_uploader" name="img_uploader" accept="image/*" />
                                 </div>
                                 <hr class="card_divider">
             
                                 <div class="user_info">
-                                    <h3><span class='capitalize lastname'></span> <span class='capitalize firstname'></span> <i class="fas fa-check-circle"></i></h3>
+                                    <h3>
+                                        <span id="user_info_firstname" class='capitalize lastname'></span>
+                                        <span id="user_info_lastname" class='capitalize firstname'></span>
+                                        <i class="fas fa-check-circle"></i>
+                                    </h3>
                                     <p><span id="email"></span></p>
                                     <p><span id="phone"></span></p>
                                 </div>
@@ -158,7 +158,12 @@
                 </div>
             </footer>
         </div>
+
         <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-        <script src="./js/dashboard.js"></script>
+        <script type="module" src="./js/common/user_loans.js"></script>
+        <script type="module" src="./js/common/logged_in_menu.js"></script>
+        <script type="module" src="./js/common/user_card_info.js"></script>
+        <script type="module" src="./js/common/user_card_address.js"></script>
+        <script type="module" src="./js/dashboard.js"></script>
     </body>
 </html>
