+++ UI/js/common/user.js
@@ -1,0 +1,12 @@
+const user = JSON.parse(localStorage.user);
+
+export const { id } = user;
+export const { email } = user;
+export const { firstname } = user;
+export const { lastname } = user;
+export const { photo } = user;
+export const { phone } = user;
+export const { isadmin } = user;
+export const { status } = user;
+export const { mailverified } = user;
+export const { address } = user;
