+++ UI/js/common/utils.js
@@ -1,0 +1,5 @@
+export const Formatter = new Intl.NumberFormat('en-US', {
+    style: 'decimal',
+    currency: 'NGN',
+    minimumFractionDigits: 2
+});
