+++ UI/profile.edit.html
@@ -106,6 +106,6 @@
         </div>
 
         <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-        <script src="./js/profile_edit.js"></script>
+        <script type="module" src="./js/profile_edit.js"></script>
     </body>
 </html>
